ref: Reorganize Day 6 files
