+++ lib/puzzles/2023/day06/day06.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+require_relative "part1"
+require_relative "part2"
+
+module AdventOfCode
+  module Puzzles2023
+    ##
+    # Advent of Code 2023 - Day 6
+    # https://adventofcode.com/2023/day/6
+    # {include:file:lib/puzzles/2023/day06/README.md}
+    module Day06
+    end
+  end
+end
