+++ lib/puzzles/2023/day06/input.txt
